requirements.txt
# requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
pytest==7.4.3
pytest-django==4.5.2